Fix SproutCore dependent key chain behavior with null intermediate objects
The actual bug-fix is straightforward, once found. The dependent property chain is instantiated with a recursive method named activate(). The activate() method has a convenience API that allows the top-level client to call it without an argument and it will assume that this is the root of the chain and should be linked to the top-level object. Unfortunately, when a null or non-existent object is encountered further down the chain, this root-defaulting behavior kicks in inappropriately, which results in every link below that point being attached to the top-level object. The fix is to require that the top-level client explicitly specify the root object and then eliminate the root-defaulting behavior of the recursive activate() method. (It required fewer characters to fix the bug than to describe the fix to the bug.)

frameworks/sproutcore/frameworks/runtime/mixins/observable.js [ In registerDependentKey(), explicitly specify the activation object rather than relying on it being defaulted appropriately. ]

frameworks/sproutcore/frameworks/runtime/private/property_chain.js [ In activate(), don't default null/undefined objects to the top-level target. Clients must specify the root object explicitly. ]