- Provide mechanism to determine when completion of a request has been deferred
SproutCore has a built-in performance monitor which detects when internal record queries are taking "too long" and breaks them up into chunks which can be scheduled for incremental processing. When a query returns, it may or may not contain complete results, because further processing may have been scheduled for later. Unfortunately, SproutCore provides no obvious mechanism for determining that this is the case, so we can't immediately tell whether we have complete results or not upon completion of our request.

To remedy this, we add a new 'scheduledFlushCount' property to SproutCore's SC.RecordArray class, which is incremented when further processing is scheduled and decremented when that further processing is completed. We then use that to delay our processing until the SC.RecordArray has completed all of its deferred processing.

frameworks/datastore/system/record_array.js [ Add 'scheduledFlushCount' property, which is incremented when flush() schedules additional processing, and decremented when the scheduled processing completes. ]