- SC.View's clippingFrame accounts for scroll offsets
The 'clippingFrame' property of views is computed by intersecting the frame of the view with each of its parent views. By default this 'clippingFrame' property did not take scroll position into account when performing its intersections. We must walk the parent view hierarchy, converting each view's frame to global coordinates, which includes taking the scroll position into account, and then intersect it with the current view's frame, before finally converting back to view coordinates. Except that we need to detect the special case of the SC.ContainerView that is the immediate child of the SC.ScrollView and undo the effect of the scroll position for that one view before performing the intersection; otherwise that one view completely screws up the clipping computations.

frameworks/core_foundation/views/view.js [ Fix bug in 'clippingFrame' which didn't take scroll positions into account. ]