- Improve performance of SC.Query.compare()
Replace calls to inefficient SC.compare() with a local compareStoreKeys() function which need not be as general as SC.compare(). Replace call to SC.typeOf() with JavaScript typeof operator.

frameworks/datastore/system/query.js [ In compare() method, replace calls to inefficient SC.compare() with a local compareStoreKeys() function which need not be as general as SC.compare() and replace call to SC.typeOf() with JavaScript typeof operator. ]