[#121609065] Fix Firefox crashes.
Problem was in Sproutcore mousemove handler in the RootResponder.
In 1.11 it assumes that the event target is an Element, but Firefox can set the target as the Document.
The original source of the problem, from a CODAP perspective had to do with the presence of a data tip in the
vicinity of the mouse move. Apparently, if it is removed while it is the target of the mousemove, Firefox will set the
target as described above.