[#151849765] Fix handling of exception in run loop
Make sure that run loop is closed or killed if uncaught exception
is thrown within. Previous change handled exception in the
run-loop body, but not in runloop close. We display an error
advising the user to save their work and restart as soon as
possible. If the run-loop is killed, then observers and
notifications that were active will no longer occur. Sometimes
a component can be deleted and recreated and will be functional.